Second round of bson extraction
