--- bson2/utils_test.go
@@ -1,9 +1,0 @@
-package bson2
-
-import (
-	"math/big"
-)
-
-func b(i int) *big.Int {
-	return big.NewInt(int64(i))
-}
